fix: melhorias no código

+++ Backend/src/PopaDin.Bkd.Api/Dtos/Budget/CreateBudgetRequest.cs
@@ -3,7 +3,5 @@ namespace PopaDin.Bkd.Api.Dtos.Budget;
 public class CreateBudgetRequest
 {
     public string Name { get; set; } = "";
-    public double Goal { get; set; }
-    // public int UserId { get; set; }
-    // public DateTime FinishAt { get; set; }
+    public decimal Goal { get; set; }
 }
