+++ Backend/src/PopaDin.Bkd.Api/Dtos/User/UpdateUserRequest.cs
@@ -10,5 +10,5 @@ public class UpdateUserRequest
     [MinLength(6, ErrorMessage = "A senha deve ter no mínimo 6 caracteres")]
     public string? Password { get; set; }
 
-    public double Balance { get; set; }
+    public decimal Balance { get; set; }
 }
