+++ Backend/src/PopaDin.Bkd.Domain/Interfaces/Repositories/IBudgetRepository.cs
@@ -1,5 +1,4 @@
 using PopaDin.Bkd.Domain.Models;
-using PopaDin.Bkd.Domain.Models.Budget;
 
 namespace PopaDin.Bkd.Domain.Interfaces.Repositories;
 
@@ -7,8 +6,8 @@ public interface IBudgetRepository
 {
     Task<Budget> CreateBudgetAsync(Budget budget);
     Task<PaginatedResult<Budget>> GetBudgetsAsync(ListBudgets listBudgets);
-    Task<Budget> FindBudgetByIdAsync(decimal budgetId, decimal userId);
+    Task<Budget> FindBudgetByIdAsync(int budgetId, int userId);
     Task UpdateBudgetAsync(Budget budget);
-    Task DeleteBudgetAsync(decimal budgetId);
-    Task FinishBudgetAsync(decimal budgetId);
+    Task DeleteBudgetAsync(int budgetId);
+    Task FinishBudgetAsync(int budgetId);
 }
