+++ Backend/src/PopaDin.Bkd.Domain/Interfaces/Repositories/ITagRepository.cs
@@ -1,5 +1,4 @@
 using PopaDin.Bkd.Domain.Models;
-using PopaDin.Bkd.Domain.Models.Tag;
 
 namespace PopaDin.Bkd.Domain.Interfaces.Repositories;
 
@@ -7,8 +6,8 @@ public interface ITagRepository
 {
     Task<Tag> CreateTagAsync(Tag tag);
     Task<PaginatedResult<Tag>> GetTagsAsync(ListTags listTags);
-    Task<Tag> FindTagByIdAsync(decimal tagId, decimal userId);
-    Task<List<Tag>> FindTagsByIdsAsync(List<int> ids, decimal userId);
+    Task<Tag> FindTagByIdAsync(int tagId, int userId);
+    Task<List<Tag>> FindTagsByIdsAsync(List<int> ids, int userId);
     Task UpdateTagAsync(Tag tag);
-    Task DeleteTagAsync(decimal tagId);
+    Task DeleteTagAsync(int tagId);
 }
