+++ Backend/src/PopaDin.Bkd.Domain/Interfaces/Repositories/IUserRepository.cs
@@ -1,5 +1,4 @@
 using PopaDin.Bkd.Domain.Models;
-using PopaDin.Bkd.Domain.Models.User;
 
 namespace PopaDin.Bkd.Domain.Interfaces.Repositories;
 
@@ -7,9 +6,9 @@ public interface IUserRepository
 {
     Task<User> CreateUserAsync(User user);
     Task<PaginatedResult<User>> GetUsersAsync(ListUsers listUsers);
-    Task<User> FindUserByIdAsync(decimal userId);
+    Task<User> FindUserByIdAsync(int userId);
     Task UpdateUserAsync(User user);
-    Task DeleteUserAsync(decimal userId);
+    Task DeleteUserAsync(int userId);
     Task<User> FindUserByEmailAsync(string userEmail);
-    Task UpdateBalanceAsync(decimal userId, double amount);
+    Task UpdateBalanceAsync(int userId, decimal amount);
 }
