+++ Backend/src/PopaDin.Bkd.Domain/Interfaces/Services/IAlertService.cs
@@ -1,11 +1,11 @@
-using PopaDin.Bkd.Domain.Models.Alert;
+using PopaDin.Bkd.Domain.Models;
 
 namespace PopaDin.Bkd.Domain.Interfaces.Services;
 
 public interface IAlertService
 {
-    Task<Alert> CreateAlertAsync(Alert alert, decimal userId);
-    Task<List<Alert>> GetAlertsByUserIdAsync(decimal userId);
-    Task ToggleAlertAsync(string alertId, bool active, decimal userId);
-    Task DeleteAlertAsync(string alertId, decimal userId);
+    Task<Alert> CreateAlertAsync(Alert alert, int userId);
+    Task<List<Alert>> GetAlertsByUserIdAsync(int userId);
+    Task ToggleAlertAsync(string alertId, bool active, int userId);
+    Task DeleteAlertAsync(string alertId, int userId);
 }
