+++ PopaDin.AlertService/Models/RecordCreatedEvent.cs
@@ -3,7 +3,7 @@ namespace PopaDin.AlertService.Models;
 public class RecordCreatedEvent
 {
     public int UserId { get; set; }
-    public double Value { get; set; }
+    public decimal Value { get; set; }
     public string Operation { get; set; } = "";
-    public double NewBalance { get; set; }
+    public decimal NewBalance { get; set; }
 }
